+++ antarest/study/output/storage/parquet_output_storage.py
@@ -19,8 +19,8 @@ from typing_extensions import override
 
 from antarest.core.utils.archives import ArchiveFormat, archive_dir, extract_archive
 from antarest.study.model import MatrixFrequency, MatrixIndex, StudySimResultDTO, StudySimSettingsDTO
-from antarest.study.output.file.extract_metadata import extract_metadata
-from antarest.study.output.lfs.large_file_storage import ILargeFileStorage
+from antarest.study.output.filestudy.extract_metadata import extract_metadata
+from antarest.study.output.lfs.lfs import ILargeFileStorage
 from antarest.study.output.output_model import OutputVariablesList
 from antarest.study.output.output_storage import IOutputStorage, OutputStorageType
 from antarest.study.output.storage.repository import OutputMetadata, OutputMetadataRepository
